Interactivity API: Router's experimental full-page script module should not be bundled in WordPress.
The `"./full-page"` entry defined in the `@wordpress/interactivity-router`'s `wpScriptModuleExport` field should not be bundled in WordPress Core yet as it is experimental. There is no mechanism in `wpScriptModuleExports` right now to support building a script module but not including it in WordPress Core. Since this is likely to be a rare instance we have opted to hard-code an exclusion in WordPress Core's build until this module is no longer experimental.

Props samueljseay, westonruter, youknowriad.
Fixes #64137.

git-svn-id: https://develop.svn.wordpress.org/trunk@61072 602fd350-edb4-49c9-b593-d223f7449a82