Block editor: Don't load fit text frontend scripts on the editor.
This commit fixes an issue where the FitText frontend script (fit-text-frontend.js) was incorrectly being enqueued and loaded in the block editor when editing posts that contain blocks with the fitText attribute enabled. It had no big impact has the script does not do anything on the editor (the classes the script searches are not there) but still it's useless bytes we are sending.

Developed in #10437.

Props jorgefilipecosta, oandregal.
Fixes #64173.

git-svn-id: https://develop.svn.wordpress.org/trunk@61115 602fd350-edb4-49c9-b593-d223f7449a82