Privacy: Set `$request->confirm_key` earlier so it can be reused in 2 places.
This commit updates the `wp_send_user_request()` function so that the updated confirmation key of the request is  available to both the confirmation URL and the subsequent filters (specifically `user_request_action_email_content`).

This maintains the existing behavior of generating a new key just-in-time before every user request email is sent.

Props birgire, dingo_d, garrett-eclipse, johnjamesjacoby.

Fixes #44940.

git-svn-id: https://develop.svn.wordpress.org/trunk@61137 602fd350-edb4-49c9-b593-d223f7449a82