Code Modernization: Fix instances of using `null` as an array offset.
This avoids using multisite-specific tables and columns in `WP_Date_Query::validate_column()` when on a non-multisite install. Attempting to use `$wpdb->blogs` as an array index when `null` on a non-multisite install causes a deprecation notice in PHP 8.5. This also updates the multisite table alias in `wpdb` to make it clear they could be `null`.

Developed in https://github.com/WordPress/wordpress-develop/pull/10498

Follow-up to [60904], [60809], [37477].

Props westonruter, desrosj.
See #63061.
Fixes #63957.


git-svn-id: https://develop.svn.wordpress.org/trunk@61191 602fd350-edb4-49c9-b593-d223f7449a82