Date/Time: Revert [60942].
The changes in [60942] aimed to add an indicator next to the default date and time formats for the site’s locale in General Settings. However, the changes do not take into account the fact that a user may have a different language than the site’s default selected as their preference.

Props swissspidy, johnbillion, SergeyBiryukov, pbearne, desrosj, wildworks.
See #64102.

git-svn-id: https://develop.svn.wordpress.org/trunk@61193 602fd350-edb4-49c9-b593-d223f7449a82