Build/Test Tools: Introduce `post-branching` Grunt task.
In [59673], a Grunt subtask was introduced to convert workflow file references in GitHub Action workflows from local ones to remote ones that target the `trunk` branch.

The `workflow-references-local-to-remote` Grunt task should be run after creating a new numbered branch to take advantage of the work started in [58165].

This commit introduces a new target for the `clean` Grunt task: `workflows`. Running `grunt clean:workflows` will delete all workflow files that are not intended to exist outside of `trunk`.

Another Grunt task (`post-branching`) has also been added as a way to group all tasks taht should run within a newly created branch. For now, this contains `clean:workflows` and `replace:workflow-references-local-to-remote`. But more can be added in the future as more aspects of the branching process are automated.

Props johnbillion.
See #64227.

git-svn-id: https://develop.svn.wordpress.org/trunk@61286 602fd350-edb4-49c9-b593-d223f7449a82