Upgrade/Install: Initialize plugin dependencies during ajax plugin search.
This ensures that a plugin's "Deactivate" link is disabled for dependency plugins during searches when a dependent plugin is active. It also ensures the "Requires" metadata is shown for dependent plugins, and the "Required by" metadata is shown for dependency plugins. 

Developed in https://github.com/WordPress/wordpress-develop/pull/10547

Follow-up to [57545].

Props hbhalodia, webdados, swissspidy, ravichudasama01.
See #22316.
Fixes #64288.


git-svn-id: https://develop.svn.wordpress.org/trunk@61312 602fd350-edb4-49c9-b593-d223f7449a82